1000 lines is too much
