+++ packages/houdini/src/generators/runtime/template/cache/index.ts
@@ -1,0 +1,10 @@
+import { Cache } from './cache'
+
+const localCache = new Cache()
+
+if (global.window) {
+	// @ts-ignore
+	window.cache = localCache
+}
+
+export default localCache
